refactor: add app dto library
