--- project/apps/users/src/app/notify/notify.constant.ts
@@ -1,3 +1,0 @@
-export const EMAIL_NOT_VALID = 'The email is not valid';
-export const FIRST_NAME_IS_EMPTY = 'The first name is empty';
-export const USER_ID_IS_EMPTY = 'The userId is empty';
