+++ project/libs/shared/app/dto/README.md
@@ -1,0 +1,7 @@
+# dto
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test dto` to execute the unit tests via [Jest](https://jestjs.io).
