+++ project/libs/shared/app/dto/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/create-subscriber/create-subscriber.dto';
